model/user_test.go
package model_test

import (
	"context"
	"github.com/alexedwards/argon2id"
	"github.com/stretchr/testify/assert"
	"go.mongodb.org/mongo-driver/bson"
	"go.mongodb.org/mongo-driver/mongo"
	"go.mongodb.org/mongo-driver/mongo/options"
	"os"
	"server/model"
	"testing"
)

func TestCreateUsersCollection(t *testing.T) {
	ctx := context.TODO()
	b, err := os.ReadFile("./user.json")
	assert.NoError(t, err)
	var jsonSchema bson.D
	err = bson.UnmarshalExtJSON(b, true, &jsonSchema)
	assert.NoError(t, err)

	n, err := db.ListCollectionNames(ctx, bson.M{"name": "users"})
	assert.NoError(t, err)
	if len(n) == 0 {
		option := options.CreateCollection().SetValidator(jsonSchema)
		err = db.CreateCollection(ctx, "users", option)
		assert.NoError(t, err)
	} else {
		err = db.RunCommand(ctx, bson.D{
			{"collMod", "users"},
			{"validator", jsonSchema},
			{"validationLevel", "strict"},
		}).Err()
		assert.NoError(t, err)
	}

	index := mongo.IndexModel{
		Keys: bson.D{{"email", 1}},
		Options: options.Index().
			SetUnique(true).
			SetName("idx_email"),
	}
	r, err := db.Collection("users").Indexes().CreateOne(ctx, index)
	assert.NoError(t, err)
	t.Log(r)
}

func TestCreateUser(t *testing.T) {
	hash, err := argon2id.CreateHash(
		"pass@VAN1234",
		argon2id.DefaultParams,
	)
	assert.NoError(t, err)
	_, err = db.Collection("users").InsertOne(
		context.TODO(),
		model.NewUser("zhangtqx@vip.qq.com", hash),
	)
	assert.NoError(t, err)
}
